cleanup, decent state
Cleaned up the code significantly. Improvements towards using the version of the sky shader in the official A-Frame examples repo, which is also what the latest version of the A-Frame environment component uses, and so seems like the most 'official' version out there, perhaps?

Some schema attributes still need to be updated and implemented and tested.